Working json demonstration
SOme dummy experiment to start using json for communication.
The client currently sends a json-encoded message with keys "message"
and "repeat_times" to the server, and the server parses the json and
responds with the message string repeated as many times as requested.

Damn, why didn't I start using boost before? It's awesome!