Moved some genenerally useful code to common file
This is to avoid code duplication between server and client. Both will
include "../common/common.h" and when a function is useful to both, we
move it there.

I have updated the server build script accordigly, but not the client
script yet. And I have not updated the windows build scripts either.